Split trace audit modules
