+++ src/commands/trace/tests.rs
@@ -1,208 +1,14 @@
-use std::collections::HashMap;
 use std::fs;
 
 use crate::test_support::with_isolated_home;
 
-use homeboy::core::component::ScopedExtensionConfig;
-use homeboy::core::rig::ComponentSpec;
-
-use super::matrix::{expand_trace_matrix, matrix_cell_env, parse_trace_matrix_axis};
 use super::test_fixture::{
     init_overlay_component, write_missing_trace_artifact_extension,
     write_nested_trace_artifact_extension, write_trace_extension, write_trace_rig,
     write_trace_rig_with_phase_preset, write_trace_rig_with_span_metadata,
-    write_trace_rig_with_variant, TRACE_FIXTURE_EXTENSION_ID,
+    write_trace_rig_with_variant,
 };
 use super::*;
-
-fn write_matrix_sensitive_trace_extension(home: &tempfile::TempDir) {
-    write_trace_extension(home);
-    let script_path = home
-        .path()
-        .join(".config")
-        .join("homeboy")
-        .join("extensions")
-        .join(TRACE_FIXTURE_EXTENSION_ID)
-        .join("trace-runner.sh");
-    fs::write(
-        &script_path,
-        r#"#!/bin/sh
-set -eu
-if [ "$HOMEBOY_TRACE_LIST_ONLY" = "1" ]; then
-  cat > "$HOMEBOY_TRACE_RESULTS_FILE" <<JSON
-{"component_id":"$HOMEBOY_COMPONENT_ID","scenarios":[{"id":"matrix-check","source":"fixture"}]}
-JSON
-  exit 0
-fi
-if [ "${HOMEBOY_TRACE_MATRIX_VIEWPORT:-}" = "mobile" ]; then
-  status="fail"
-else
-  status="pass"
-fi
-mkdir -p "$HOMEBOY_TRACE_ARTIFACT_DIR"
-printf 'viewport=%s\nlocation=%s\n' "${HOMEBOY_TRACE_MATRIX_VIEWPORT:-}" "${HOMEBOY_TRACE_MATRIX_ECE_LOCATIONS:-}" > "$HOMEBOY_TRACE_ARTIFACT_DIR/matrix-env.txt"
-cat > "$HOMEBOY_TRACE_RESULTS_FILE" <<JSON
-{"component_id":"$HOMEBOY_COMPONENT_ID","scenario_id":"$HOMEBOY_TRACE_SCENARIO","status":"$status","timeline":[{"t_ms":0,"source":"runner","event":"boot"},{"t_ms":25,"source":"runner","event":"ready"}],"span_results":[{"id":"boot_to_ready","from":"runner.boot","to":"runner.ready","status":"ok","duration_ms":25,"from_t_ms":0,"to_t_ms":25}],"assertions":[],"artifacts":[{"label":"matrix env","path":"artifacts/matrix-env.txt"}]}
-JSON
-"#,
-    )
-    .expect("write matrix trace script");
-
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        let mut permissions = fs::metadata(&script_path)
-            .expect("script metadata")
-            .permissions();
-        permissions.set_mode(0o755);
-        fs::set_permissions(&script_path, permissions).expect("chmod script");
-    }
-}
-
-#[test]
-fn trace_matrix_expands_axes_and_exports_env() {
-    let axes = vec![
-        parse_trace_matrix_axis("viewport=desktop,mobile").expect("viewport axis"),
-        parse_trace_matrix_axis("ece_locations=product,none").expect("location axis"),
-    ];
-
-    let cells = expand_trace_matrix(&axes);
-
-    assert_eq!(cells.len(), 4);
-    assert_eq!(cells[0].label, "ece_locations-product__viewport-desktop");
-    assert_eq!(cells[3].values.get("viewport"), Some(&"mobile".to_string()));
-    assert_eq!(
-        matrix_cell_env(&cells[0])
-            .into_iter()
-            .find(|(key, _)| key == "HOMEBOY_TRACE_MATRIX_ECE_LOCATIONS")
-            .map(|(_, value)| value),
-        Some("product".to_string())
-    );
-}
-
-#[test]
-fn trace_matrix_reports_failed_cells_and_cell_artifacts() {
-    with_isolated_home(|home| {
-        write_matrix_sensitive_trace_extension(home);
-        let component_dir = tempfile::TempDir::new().expect("component dir");
-        write_trace_rig(home, "studio-rig", "studio", component_dir.path());
-        let output_dir = tempfile::TempDir::new().expect("matrix output dir");
-
-        let ((output, _artifact_output), exit_code) = run_outputs(TraceArgs {
-            comp: PositionalComponentArgs {
-                component: Some("matrix".to_string()),
-                path: None,
-            },
-            component_arg: None,
-            scenario: Some("studio".to_string()),
-            scenario_arg: None,
-            compare_after: Some("matrix-check".into()),
-            baseline_target: None,
-            candidate: None,
-            rig: Some("studio-rig".to_string()),
-            profile: None,
-            profiles: false,
-            setting_args: SettingArgs::default(),
-            json_summary: false,
-            report: None,
-            experiment: None,
-            repeat: 1,
-            aggregate: None,
-            schedule: TraceSchedule::Grouped,
-            focus_spans: Vec::new(),
-            spans: Vec::new(),
-            phases: Vec::new(),
-            attachments: Vec::new(),
-            phase_preset: None,
-            baseline_args: BaselineArgs::default(),
-            regression_threshold: extension_trace::baseline::DEFAULT_REGRESSION_THRESHOLD_PERCENT,
-            regression_min_delta_ms: extension_trace::baseline::DEFAULT_REGRESSION_MIN_DELTA_MS,
-            overlays: Vec::new(),
-            variants: Vec::new(),
-            matrix: TraceVariantMatrixMode::None,
-            axes: vec![
-                parse_trace_matrix_axis("viewport=desktop,mobile").expect("viewport axis"),
-                parse_trace_matrix_axis("ece_locations=product").expect("location axis"),
-            ],
-            matrix_env: Vec::new(),
-            output_dir: Some(output_dir.path().to_path_buf()),
-            keep_overlay: false,
-            stale: false,
-            force: false,
-        })
-        .expect("matrix run");
-
-        assert_eq!(exit_code, 1);
-        let TraceCommandOutput::ScenarioMatrix(matrix) = output else {
-            panic!("expected scenario matrix output");
-        };
-        assert_eq!(matrix.command, "trace.matrix");
-        assert_eq!(matrix.cell_count, 2);
-        assert_eq!(matrix.failure_count, 1);
-        assert!(std::path::Path::new(&matrix.matrix_path).exists());
-        assert!(std::path::Path::new(&matrix.summary_path).exists());
-        let failed = matrix
-            .cells
-            .iter()
-            .find(|cell| !cell.passed)
-            .expect("failed mobile cell");
-        assert_eq!(failed.axes.get("viewport"), Some(&"mobile".to_string()));
-        assert_eq!(failed.status, "fail");
-        assert_eq!(failed.exit_code, 1);
-        assert!(std::path::Path::new(&failed.output_path).exists());
-        assert!(std::path::Path::new(&failed.artifact_path).exists());
-        assert!(std::path::Path::new(&failed.artifact_dir)
-            .join("matrix-env.txt")
-            .exists());
-
-        let summary = fs::read_to_string(matrix.summary_path).expect("summary");
-        assert!(summary.contains("| `ece_locations-product__viewport-mobile` |"));
-        assert!(summary.contains("`fail`"));
-    });
-}
-
-fn trace_args_for_profile(profile: &str) -> TraceArgs {
-    TraceArgs {
-        comp: PositionalComponentArgs {
-            component: None,
-            path: None,
-        },
-        component_arg: None,
-        scenario: None,
-        scenario_arg: None,
-        compare_after: None,
-        baseline_target: None,
-        candidate: None,
-        rig: None,
-        profile: Some(profile.to_string()),
-        profiles: false,
-        setting_args: SettingArgs::default(),
-        json_summary: false,
-        report: None,
-        experiment: None,
-        repeat: 1,
-        aggregate: None,
-        schedule: TraceSchedule::Grouped,
-        focus_spans: Vec::new(),
-        spans: Vec::new(),
-        phases: Vec::new(),
-        attachments: Vec::new(),
-        phase_preset: None,
-        baseline_args: BaselineArgs::default(),
-        regression_threshold: extension_trace::baseline::DEFAULT_REGRESSION_THRESHOLD_PERCENT,
-        regression_min_delta_ms: extension_trace::baseline::DEFAULT_REGRESSION_MIN_DELTA_MS,
-        overlays: Vec::new(),
-        variants: Vec::new(),
-        matrix: TraceVariantMatrixMode::None,
-        axes: Vec::new(),
-        matrix_env: Vec::new(),
-        output_dir: None,
-        keep_overlay: false,
-        stale: false,
-        force: false,
-    }
-}
-
 fn trace_args_for_rig(rig_id: &str, component_id: &str, scenario_id: &str) -> TraceArgs {
     TraceArgs {
         comp: PositionalComponentArgs {
@@ -245,312 +51,6 @@ fn trace_args_for_rig(rig_id: &str, component_id: &str, scenario_id: &str) -> Tr
         canonical: false,
         allow_local_toolchain: false,
     }
-}
-
-fn write_trace_rig_with_profile(
-    home: &tempfile::TempDir,
-    rig_id: &str,
-    component_id: &str,
-    path: &std::path::Path,
-) {
-    let rig_dir = home.path().join(".config").join("homeboy").join("rigs");
-    fs::create_dir_all(&rig_dir).expect("mkdir rigs");
-    fs::write(
-        rig_dir.join(format!("{}.json", rig_id)),
-        format!(
-            r#"{{
-                    "components": {{
-                        "{component_id}": {{ "path": "{}" }}
-                    }},
-                    "trace_workloads": {{ "{TRACE_FIXTURE_EXTENSION_ID}": [
-                        {{ "path": "${{components.{component_id}.path}}/close-window-running-site.trace.mjs" }}
-                    ] }},
-                    "trace_profiles": {{
-                        "studio-window-close": {{
-                            "component": "{component_id}",
-                            "scenario": "close-window-running-site",
-                            "settings": {{
-                                "window_title": "Studio",
-                                "retry_count": 2
-                            }}
-                        }}
-                    }}
-                }}"#,
-            path.display()
-        ),
-    )
-    .expect("write rig");
-}
-
-#[test]
-fn rig_component_path_and_trace_env_are_threaded() {
-    let component_dir = tempfile::TempDir::new().expect("component dir");
-    let mut components = HashMap::new();
-    let mut extensions = HashMap::new();
-    extensions.insert(
-        "trace-extension".to_string(),
-        ScopedExtensionConfig::default(),
-    );
-    components.insert(
-        "studio".to_string(),
-        ComponentSpec {
-            path: component_dir.path().to_string_lossy().to_string(),
-            checkout_root: None,
-            remote_url: Some("https://github.com/Automattic/studio".to_string()),
-            triage_remote_url: None,
-            stack: None,
-            branch: None,
-            extensions: Some(extensions),
-        },
-    );
-    let spec = RigSpec {
-        id: "studio-rig".to_string(),
-        components,
-        ..serde_json::from_str(r#"{"id":"studio-rig"}"#).unwrap()
-    };
-    let path = rig_component_path(&spec, "studio").expect("path resolves");
-    assert_eq!(path, component_dir.path().to_string_lossy());
-    let component = rig_component_for_trace(&spec, "studio").expect("component resolves");
-    assert_eq!(component.id, "studio");
-    assert_eq!(component.local_path, path);
-    assert!(component.extensions.is_some());
-}
-
-#[test]
-fn rig_component_for_trace_synthesizes_trace_workload_extensions() {
-    let rig_spec: RigSpec = serde_json::from_str(
-        r#"{
-                "id": "studio",
-                "components": {
-                    "studio": { "path": "/tmp/studio" }
-                },
-                "trace_workloads": {
-                    "trace-fixture": [{ "path": "/tmp/create-site.trace.mjs" }]
-                }
-            }"#,
-    )
-    .expect("parse rig spec");
-
-    let component = rig_component_for_trace(&rig_spec, "studio").expect("component");
-
-    assert!(component
-        .extensions
-        .as_ref()
-        .expect("extensions")
-        .contains_key(TRACE_FIXTURE_EXTENSION_ID));
-}
-
-#[test]
-fn rig_trace_list_uses_rig_default_component_and_workloads() {
-    let component_dir = tempfile::TempDir::new().expect("component dir");
-    let rig_spec: RigSpec = serde_json::from_str(&format!(
-        r#"{{
-            "id": "studio-rig",
-            "components": {{
-                "studio": {{ "path": "{}" }}
-            }},
-            "trace_workloads": {{ "{TRACE_FIXTURE_EXTENSION_ID}": [
-                {{ "path": "${{components.studio.path}}/studio-app-create-site.trace.mjs" }},
-                {{ "path": "${{components.studio.path}}/studio-list-sites.trace.mjs" }}
-            ] }}
-        }}"#,
-        component_dir.path().display()
-    ))
-    .expect("parse rig");
-
-    let component_id = resolve_component_id(
-        &PositionalComponentArgs {
-            component: None,
-            path: None,
-        },
-        Some(&rig_spec),
-    )
-    .expect("default component");
-    let component = rig_component_for_trace(&rig_spec, &component_id).expect("component");
-    let workloads = rig::workloads_for_extension(
-        &rig_spec,
-        rig::RigWorkloadKind::Trace,
-        None,
-        TRACE_FIXTURE_EXTENSION_ID,
-    );
-    let scenarios = workloads
-        .iter()
-        .map(|path| trace_workload_scenario_id(&path.to_string_lossy()))
-        .collect::<Vec<_>>();
-
-    assert_eq!(component_id, "studio");
-    assert_eq!(component.id, "studio");
-    assert_eq!(workloads.len(), 2);
-    assert_eq!(scenarios[0], "studio-app-create-site");
-    let expected_source = format!(
-        "{}/studio-app-create-site.trace.mjs",
-        component_dir.path().display()
-    );
-    assert_eq!(workloads[0].to_string_lossy(), expected_source);
-}
-
-#[test]
-fn trace_profile_resolves_to_normal_trace_run() {
-    with_isolated_home(|home| {
-        write_trace_extension(home);
-        let component_dir = tempfile::TempDir::new().expect("component dir");
-        write_trace_rig_with_profile(home, "studio-rig", "studio", component_dir.path());
-
-        let ((output, _artifact_output), exit_code) =
-            run_outputs(trace_args_for_profile("studio-window-close"))
-                .expect("profile trace should run");
-
-        assert_eq!(exit_code, 0);
-        let TraceCommandOutput::Run(run) = output else {
-            panic!("expected run output");
-        };
-        let profile = run.profile.expect("resolved profile metadata");
-        assert_eq!(profile.id, "studio-window-close");
-        assert_eq!(profile.rig_id.as_deref(), Some("studio-rig"));
-        assert_eq!(profile.component, "studio");
-        assert_eq!(profile.scenario, "close-window-running-site");
-        assert_eq!(
-            profile.settings.get("window_title"),
-            Some(&serde_json::Value::String("Studio".to_string()))
-        );
-        assert_eq!(
-            profile.settings.get("retry_count"),
-            Some(&serde_json::json!(2))
-        );
-    });
-}
-
-#[test]
-fn trace_profile_cli_fields_override_profile_fields() {
-    with_isolated_home(|home| {
-        let component_dir = tempfile::TempDir::new().expect("component dir");
-        write_trace_rig_with_profile(home, "studio-rig", "studio", component_dir.path());
-        let mut args = trace_args_for_profile("studio-window-close");
-        args.scenario = Some("close-window-retry".to_string());
-        args.setting_args
-            .setting
-            .push(("window_title".to_string(), "Studio Dev".to_string()));
-
-        resolve_trace_profile_args(&mut args).expect("profile resolves");
-
-        assert_eq!(args.comp.component.as_deref(), Some("studio"));
-        assert_eq!(args.scenario.as_deref(), Some("close-window-retry"));
-        assert_eq!(args.rig.as_deref(), Some("studio-rig"));
-        assert_eq!(
-            args.setting_args.setting,
-            vec![
-                ("window_title".to_string(), "Studio".to_string()),
-                ("window_title".to_string(), "Studio Dev".to_string())
-            ]
-        );
-    });
-}
-
-#[test]
-fn trace_list_profiles_lists_rig_profiles() {
-    with_isolated_home(|home| {
-        let component_dir = tempfile::TempDir::new().expect("component dir");
-        write_trace_rig_with_profile(home, "studio-rig", "studio", component_dir.path());
-
-        let ((output, _artifact_output), exit_code) = run_outputs(TraceArgs {
-            comp: PositionalComponentArgs {
-                component: Some("list".to_string()),
-                path: None,
-            },
-            component_arg: None,
-            scenario: None,
-            scenario_arg: None,
-            compare_after: None,
-            baseline_target: None,
-            candidate: None,
-            rig: None,
-            profile: None,
-            profiles: true,
-            setting_args: SettingArgs::default(),
-            json_summary: false,
-            report: None,
-            experiment: None,
-            repeat: 1,
-            aggregate: None,
-            schedule: TraceSchedule::Grouped,
-            focus_spans: Vec::new(),
-            spans: Vec::new(),
-            phases: Vec::new(),
-            attachments: Vec::new(),
-            phase_preset: None,
-            baseline_args: BaselineArgs::default(),
-            regression_threshold: extension_trace::baseline::DEFAULT_REGRESSION_THRESHOLD_PERCENT,
-            regression_min_delta_ms: extension_trace::baseline::DEFAULT_REGRESSION_MIN_DELTA_MS,
-            overlays: Vec::new(),
-            variants: Vec::new(),
-            matrix: TraceVariantMatrixMode::None,
-            axes: Vec::new(),
-            matrix_env: Vec::new(),
-            output_dir: None,
-            keep_overlay: false,
-            stale: false,
-            force: false,
-        })
-        .expect("profile list should run");
-
-        assert_eq!(exit_code, 0);
-        let TraceCommandOutput::List(list) = output else {
-            panic!("expected list output");
-        };
-        assert_eq!(list.command, "trace.list.profiles");
-        assert_eq!(list.count, 1);
-        assert_eq!(list.profiles[0].id, "studio-window-close");
-        assert_eq!(list.profiles[0].rig_id, "studio-rig");
-        assert_eq!(list.profiles[0].component.as_deref(), Some("studio"));
-        assert_eq!(
-            list.profiles[0].scenario.as_deref(),
-            Some("close-window-running-site")
-        );
-    });
-}
-
-#[test]
-fn rig_trace_list_uses_scoped_workload_preflight() {
-    let spec: RigSpec = serde_json::from_str(&format!(
-        r#"{{
-            "id": "studio-rig",
-            "components": {{
-                "studio": {{ "path": "/tmp/studio" }}
-            }},
-            "pipeline": {{
-                "check": [
-                    {{
-                        "kind": "check",
-                        "label": "desktop app packaged",
-                        "groups": ["desktop-app"],
-                        "command": "true"
-                    }},
-                    {{
-                        "kind": "check",
-                        "label": "unrelated cli symlink",
-                        "groups": ["cli-dev-copy"],
-                        "command": "false"
-                    }}
-                ]
-            }},
-            "trace_workloads": {{ "{TRACE_FIXTURE_EXTENSION_ID}": [
-                {{
-                    "path": "${{components.studio.path}}/studio-app-create-site.trace.mjs",
-                    "check_groups": ["desktop-app"]
-                }}
-            ] }}
-        }}"#
-    ))
-    .expect("parse rig");
-
-    run_rig_workload_preflight(
-        &spec,
-        Some(TRACE_FIXTURE_EXTENSION_ID),
-        rig::RigWorkloadKind::Trace,
-    )
-    .expect("scoped workload preflight should bypass unrelated failed check");
-
-    assert!(run_rig_workload_preflight(&spec, None, rig::RigWorkloadKind::Trace).is_err());
 }
 
 #[test]
